requirements.txt
amqp==2.6.1
anyjson==0.3.3
billiard==3.6.3.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==2.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.7.0
django-environ==0.4.5
dnspython==2.1.0
greenlet==1.0.0
kombu==4.6.11
monotonic==1.5
prompt-toolkit==3.0.16
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
vine==1.3.0
wcwidth==0.2.5
